requirements.txt
# Core Data Science
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Data Storage
pyarrow>=12.0.0
fastparquet>=2023.4.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0

# Deep Learning (optional - for LSTM)
torch>=2.0.0

# Time Series
statsmodels>=0.14.0
prophet>=1.1.4

# Explainability
shap>=0.42.0

# API
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# Dashboard
streamlit>=1.25.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
joblib>=1.3.0
PyYAML>=6.0

# Development
jupyter>=1.0.0
jupyterlab>=4.0.0
ipykernel>=6.25.0
black>=23.7.0
isort>=5.12.0
flake8>=6.1.0
pytest>=7.4.0
pytest-cov>=4.1.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.1.0